Move interfaces para seus respectivos recursos
