+++ data_extraction/__init__.py
@@ -1,1 +1,2 @@
+from .interfaces import TextExtractorInterface
 from .text_extraction import ApacheTikaTextExtractor, create_apache_tika_text_extraction
