+++ database/__init__.py
@@ -1,1 +1,2 @@
+from .interfaces import DatabaseInterface
 from .postgresql import PostgreSQL, create_database_interface
