+++ index/__init__.py
@@ -1,1 +1,2 @@
+from .interfaces import IndexInterface
 from .opensearch import create_index_interface
