+++ storage/__init__.py
@@ -1,1 +1,2 @@
 from .digital_ocean_spaces import DigitalOceanSpaces, create_storage_interface
+from .interfaces import StorageInterface
